requirements.txt
pyTelegramBotAPI==4.12.0  # Main Telegram bot library
python-dotenv==1.0.0      # For loading .env files